clean up wallet integration
