+++ src/containers/Web3Wrapper/Web3ReactManager.tsx
@@ -1,8 +1,9 @@
 import React, { useState, useEffect } from 'react';
 import { useWeb3React } from '@web3-react/core';
-import { useEagerConnect, useInactiveListener } from '../hooks/wallet';
 
-export default function Web3ReactManager({ children }: $TSFixMe) {
+import { useEagerConnect, useInactiveListener } from 'hooks/wallet';
+
+const Web3ReactManager: React.FC = ({ children }) => {
   const context = useWeb3React();
   const {
     connector,
@@ -22,19 +23,24 @@ export default function Web3ReactManager({ children }: $TSFixMe) {
     }
   }, [activatingConnector, connector, chainId]);
 
-  // when there's no account connected, react to logins (broadly speaking) on the injected provider, if it exists
+  // when there's no account connected, react to logins (broadly speaking) on
+  // the injected provider, if it exists
   useInactiveListener(!triedEager);
 
-  // on page load, do nothing until we've tried to connect to the injected connector
+  // on page load, do nothing until we've tried to connect to the injected
+  // connector
   if (!triedEager) {
     return null;
   }
 
-  // if the account context isn't active, and there's an error on the network context, it's an irrecoverable error
+  // if the account context isn't active, and there's an error on the network
+  // context, it's an irrecoverable error
   if (!active && networkError) {
-    console.log('unknownError');
-    return <>unknownError</>;
+    // TODO: log error to Sentry
+    return <>An internal error occurred. Please try again later.</>;
   }
 
-  return children;
-}
+  return <>{children}</>;
+};
+
+export default Web3ReactManager;
