+++ src/hooks/wallet/useAuth.ts
@@ -10,11 +10,11 @@ import {
   WalletConnectConnector,
 } from '@web3-react/walletconnect-connector';
 import toast from 'components/Basic/Toast';
-import { LS_KEY_IS_USER_LOGGED_IN } from '../config';
-import { connectorsByName, ConnectorNames } from '../utilities/connectors';
-import { setupNetwork } from '../utilities/wallet';
+import { LS_KEY_IS_USER_LOGGED_IN } from '../../config';
+import { connectorsByName, ConnectorNames } from '../../utilities/connectors';
+import { setupNetwork } from '../../utilities/wallet';
 
-const useAuth = () => {
+export default function useAuth() {
   const { activate, deactivate } = useWeb3React();
   const login = useCallback(
     async (connectorID: ConnectorNames) => {
@@ -92,6 +92,4 @@ const useAuth = () => {
   }, [deactivate]);
 
   return { login, logout };
-};
-
-export default useAuth;
+}
